yaaaay u can play audio

+++ src/main/kotlin/commands/Play.kt
@@ -3,4 +3,4 @@ package commands
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Play(val guild: Double, val name: String, val cache: Boolean=true, val requester: Double, val channel: Double)
+data class Play(val guild: Long, val name: String, val cache: Boolean=true, val requester: Double, val channel: Double)
