+++ src/main/kotlin/commands/Skip.kt
@@ -3,4 +3,4 @@ package commands
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Skip(val guild: Double)
+data class Skip(val guild: Long)
